requirements.txt
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6-sip==13.8.0
Pillow>=10.0
opencv-python>=4.8
numpy>=1.24
